% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consumidor.R
\docType{data}
\name{consumidor}
\alias{consumidor}
\title{consumidor}
\format{data.frame with 2000 obs and nine columns:
\describe{
  \item{\code{processo}}{character docket number}
  \item{\code{classe}}{character type of precedure}
  \item{\code{assunto}}{character subject}
  \item{\code{magistrado}}{character judge's name}
  \item{\code{comarca}}{character judicial circunscription}
  \item{\code{foro}}{character court}
  \item{\code{vara}}{character judge's prescint}
  \item{\code{disponibilizacao}}{Date availailability date}
  \item{\code{julgado}}{character decision whole text}
}}
\source{
\url{https://tjsp.jus.br/}
}
\usage{
consumidor
}
\description{
consumers decisions sample from Sao Paulo courts
}
\keyword{datasets}
