% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_add_pkey.R
\name{psql_add_pkey}
\alias{psql_add_pkey}
\title{Adds primary key to table}
\usage{
psql_add_pkey(con, tbl)
}
\arguments{
\item{con}{connection}

\item{tbl}{table's name}
}
\value{
It doesn't return anything, it just adds
   a primary key to the table.
}
\description{
Adds primary key to table
}
\examples{
\dontrun{
data(consumidor)
con <- dbx::dbxConnect()
dplyr::copy_to(con, "consumidor", "consumidor")
psql_add_pkey(con, "consumidor")
}
}
