% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_cjpg_query.R
\name{psql_cjpg_query}
\alias{psql_cjpg_query}
\title{PostgreSQL full-text search}
\usage{
psql_cjpg_query(con, tbl, query, assuntos = NULL, foros = NULL)
}
\arguments{
\item{con}{connection}

\item{tbl}{table}

\item{query}{words to be searched. Separate by "&" (AND), "|" (OR),
or "<->" (FOLLOWED BY) to search for more than one word.}

\item{assuntos}{vector of subjects to filter. Defaults do all subjects}

\item{foros}{vectos of courts to filter. Defaults to all courts.}
}
\value{
A data.frame with the cjpg decisions according to
}
\description{
PostgreSQL full-text search
}
\examples{
\dontrun{
dplyr::copy_to(con, "consumidor", consumidor)
df <- psql_cjpg_query(con, "consumidor", "contrato & alguel")
df <- psql_cjpg_query(con, "consumidor", "contrato | alguel")
# Search for Julgo Procedente separated by up to 3 words:
df <- psql_cjpg_query(con, "consumidor", "julgo <3> procedente")
}
}
