% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_cjpg_tokenize.R
\name{psql_cjpg_tokenize}
\alias{psql_cjpg_tokenize}
\title{Creates an index from columns assunto and julgados}
\usage{
psql_cjpg_tokenize(con, tbl)
}
\arguments{
\item{con}{connection}

\item{tbl}{table}
}
\value{
No R object is returned
}
\description{
Creates an index from columns assunto and julgados
}
\examples{
\dontrun{
library(JurisPsql)
con <- dbx::dbxConnect()
dplyr::copy_to(con, "consumidor", consumidor)
psql_cjpg_tokenize(con, "consumidor")
}
}
