% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_classify_decision.R
\name{psql_classify_decision}
\alias{psql_classify_decision}
\title{Creates a new column giving a one word conclusion of the decision}
\usage{
psql_classify_decision(con, tbl = NULL, decision = NULL,
  conclusion = NULL)
}
\arguments{
\item{con}{connection}

\item{tbl}{table with the whole decision text}

\item{decision}{decision text}

\item{conclusion}{name of new column}
}
\description{
Creates a new column giving a one word conclusion of the decision
}
