% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_do.R
\name{psql_do}
\alias{psql_do}
\title{Retrives database information}
\usage{
psql_do(token = NULL)
}
\arguments{
\item{token}{you get this token from your digital ocean
account.}
}
\value{
a list with all details about the database.
}
\description{
Retrives database information
}
