% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psql_write_cjpg.R
\name{psql_write_cjpg}
\alias{psql_write_cjpg}
\title{Create and write a cjpg table to a tjsp database}
\usage{
psql_write_cjpg(con = NULL, tbl = NULL, data = NULL)
}
\arguments{
\item{con}{a connection}

\item{tbl}{table name}

\item{data}{cjpg data.frame to be written.}
}
\value{
returns NULL if everything went well
}
\description{
This function firstly creates a table based on the
    data columns to be written. Then it adds a primary key called
    document_idx. Secondly, it inserts the data rows. After that it
    creates an inverted index called document_tokens based on GIN algo.
}
\examples{
\dontrun{
con <- dbx::dbxConnect()
psql(con, "consumidor", consumidor)
}
}
